Fixes problem with multicheckbox in more than one form on same page
More forms on single page with one multicheckbox per form with same
relation. Results in deleting relation-rows in submited form (no
problem with the first form on page) after changing values.
It has todo with postData-variable names. On first request every form
is initialised and therefore multicheckbox#1 is form_RELATION_ID,
multicheckbox in form2 is form_RELATION2_ID and so on. In post-request
only posted form is initialised and therefore multicheckbox in form#2
is called form_RELATION_ID and does not match with post-variables
with value form_RELATION2_ID.